Constraints can now be added to the config file
As of now only one type of constraint is supported: end of day
constraint. This constraint can check at the end of each day whether the
balance in an account has the correct sign.

For example, a cash account can be ensured to be always positive (or
always negative, depending on the accounting convention).