File level scopes. This changes how "include" directives are processed
For GH-5

All files have an implicit scope.

When a ledger file (F) is included in a scope(S), a new scope (L) is created for the file with S as the parent scope.
The entries in L are merged with S, such that entries in S can refer to definitions from L.
And entries in L can refer to definitions from S.