consistent export types and formats
earlier, exporting with the journal type and ledger format resulted in a
single entry containing the balance of all accounts.

With this change, the export types and formats are less confusing:

Export type determines whether a summary (balance) is exported or a
journal of all transactions is exported.

Export format determines whether the output is in ledger format or xml
format.